+++ Modules/Language/translate.c
@@ -2,6 +2,9 @@
 
 #include "translate.h"
 
+#include <stdio.h>
+#include <string.h>
+
 #include "hal_defs.h"
 
 
@@ -69,6 +72,46 @@ const char T_RIGHT[][TRANSLATE_MAX_LEN] = {
 	"RIGHT",
 	"¬œ–¿¬Œ"
 };
+const char T_UNKNOWN_ERROR[][TRANSLATE_MAX_LEN] = {
+	"UNKNOWN ERROR",
+	"Õ≈»«¬≈—“Õ¿ﬂ Œÿ»¡ ¿"
+};
+const char T_INTERNAL_ERROR[][TRANSLATE_MAX_LEN] = {
+	"INTERNAL ERROR",
+	"¬Õ”“–≈ÕÕﬂﬂ Œÿ»¡ ¿"
+};
+const char T_MEMORY_ERROR[][TRANSLATE_MAX_LEN] = {
+	"MEMORY ERROR",
+	"Œÿ»¡ ¿ œ¿Ãﬂ“»"
+};
+const char T_POWER_ERROR[][TRANSLATE_MAX_LEN] = {
+	"POWER ERROR",
+	"Œÿ»¡ ¿ œ»“¿Õ»ﬂ"
+};
+const char T_STACK_ERROR[][TRANSLATE_MAX_LEN] = {
+	"STACK ERROR",
+	"Œÿ»¡ ¿ —“≈ ¿"
+};
+const char T_LOAD_ERROR[][TRANSLATE_MAX_LEN] = {
+	"LOAD ERROR",
+	"Œÿ»¡ ¿ «¿√–”« »"
+};
+const char T_RAM_ERROR[][TRANSLATE_MAX_LEN] = {
+	"RAM ERROR",
+	"Œÿ»¡ ¿ Œ«”"
+};
+const char T_SETTINGS_ERROR[][TRANSLATE_MAX_LEN] = {
+	"SETTINGS ERROR",
+	"Œÿ»¡ ¿ Õ¿—“–Œ≈ "
+};
+const char T_APP_ERROR[][TRANSLATE_MAX_LEN] = {
+	"APPLICATION ERROR",
+	"Œÿ»¡ ¿ œ–»ÀŒ∆≈Õ»ﬂ"
+};
+const char T_VALVE_ERROR[][TRANSLATE_MAX_LEN] = {
+	"VALVE ERROR",
+	"Œÿ»¡ ¿  À¿œ¿Õ¿"
+};
 
 
 const char* t(const char phrase[][TRANSLATE_MAX_LEN], uint8_t lang)
@@ -76,4 +119,68 @@ const char* t(const char phrase[][TRANSLATE_MAX_LEN], uint8_t lang)
 	assert_param(IS_LANGUAGE(lang));
 
 	return (const char*)phrase[lang];
+};
+
+const char* get_string_error(SOUL_STATUS error, uint8_t lang)
+{
+	static char error_line[2*TRANSLATE_MAX_LEN + 2] = "";
+	memset(error_line, 0, sizeof(error_line));
+	snprintf(
+		error_line,
+		sizeof(error_line) - 1,
+		"%s",
+		t(T_UNKNOWN_ERROR, lang)
+	);
+
+	if (error <= ERRORS_START && error >= ERRORS_END) {
+		return error_line;
+	};
+
+	const char* error_ptr = NULL;
+	switch (error)
+	{
+	case INTERNAL_ERROR:
+		error_ptr = T_INTERNAL_ERROR;
+		break;
+	case MEMORY_ERROR:
+		error_ptr = T_MEMORY_ERROR;
+		break;
+	case POWER_ERROR:
+		error_ptr = T_POWER_ERROR;
+		break;
+	case STACK_ERROR:
+		error_ptr = T_STACK_ERROR;
+		break;
+	case LOAD_ERROR:
+		error_ptr = T_LOAD_ERROR;
+		break;
+	case RAM_ERROR:
+		error_ptr = T_RAM_ERROR;
+		break;
+	case USB_ERROR:
+		error_ptr = T_UNKNOWN_ERROR;
+		break;
+	case SETTINGS_LOAD_ERROR:
+		error_ptr = T_LOAD_ERROR;
+		break;
+	case APP_MODE_ERROR:
+		error_ptr = T_APP_ERROR;
+		break;
+	case VALVE_ERROR:
+		error_ptr = T_VALVE_ERROR;
+		break;
+	default:
+		error_ptr = T_UNKNOWN_ERROR;
+		break;
+	};
+
+	memset(error_line, 0, sizeof(error_line));
+	snprintf(
+		error_line,
+		sizeof(error_line) - 1,
+		"%s",
+		t(error_ptr, lang)
+	);
+
+	return error_line;
 }