+++ Modules/SettingsDB/SettingsDB.cpp
@@ -27,7 +27,7 @@ SettingsStatus SettingsDB::load()
     status = storage->find(FIND_MODE_EQUAL, &address1, PREFIX, 1);
     if (status != STORAGE_OK) {
 #if SETTINGS_BEDUG
-        printTagLog(SettingsDB::TAG, "error load settings: try to find duplicate (error=%02X)", status);
+        printTagLog(SettingsDB::TAG, "error load original settings (%02X)", status);
 #endif
         needResaveFirst = true;
     }
@@ -35,7 +35,7 @@ SettingsStatus SettingsDB::load()
     status = storage->find(FIND_MODE_EQUAL, &address2, PREFIX, 2);
     if (status != STORAGE_OK) {
 #if SETTINGS_BEDUG
-        printTagLog(SettingsDB::TAG, "error load settings: storage find error=%02X", status);
+        printTagLog(SettingsDB::TAG, "error load duplicate settings (%02X)", status);
 #endif
         needResaveSecond = true;
     }
@@ -58,7 +58,7 @@ SettingsStatus SettingsDB::load()
     memcpy(this->settings, &tmpSettings, this->size);
 
 #if SETTINGS_BEDUG
-    printTagLog(SettingsDB::TAG, "settings loaded");
+    printTagLog(SettingsDB::TAG, "load settings: success, address1=%lu, address2=%lu", address1, address2);
 #endif
 
     if (needResaveFirst || needResaveSecond) {
@@ -76,17 +76,16 @@ SettingsStatus SettingsDB::save()
 	StorageStatus status = STORAGE_OK;
 
     status = storage->find(FIND_MODE_EQUAL, &address, PREFIX, 1);
-
     if (status != STORAGE_OK) {
 #if SETTINGS_BEDUG
-        printTagLog(SettingsDB::TAG, "error save settings: storage find error, try to find duplicate (error=%02X)", status);
+        printTagLog(SettingsDB::TAG, "error save settings: storage find error (%02X), try to find duplicate", status);
 #endif
     	status = storage->find(FIND_MODE_EQUAL, &address, PREFIX, 2);
     }
 
     if (status == STORAGE_NOT_FOUND) {
 #if SETTINGS_BEDUG
-        printTagLog(SettingsDB::TAG, "error save settings: storage find duplicate error, try to find empty (error=%02X)", status);
+        printTagLog(SettingsDB::TAG, "error save settings: storage find duplicate error (%02X), try to find empty", status);
 #endif
         status = storage->find(FIND_MODE_EMPTY, &address);
     }
@@ -113,31 +112,38 @@ SettingsStatus SettingsDB::save()
         printTagLog(SettingsDB::TAG, "error save settings: storage save error=%02X address=%lu", status, address);
 #endif
         return SETTINGS_ERROR;
+    } else {
+#if SETTINGS_BEDUG
+        printTagLog(SettingsDB::TAG, "save original settings: success address=%lu", address);
+#endif
     }
 
     // Save duplicate settings
 	status = storage->find(FIND_MODE_EQUAL, &address, PREFIX, 2);
-
     if (status == STORAGE_NOT_FOUND) {
 #if SETTINGS_BEDUG
         printTagLog(SettingsDB::TAG, "error save settings duplicate: storage find error, try to find empty (error=%02X)", status);
 #endif
         status = storage->find(FIND_MODE_EMPTY, &address);
     }
-
-	status = storage->rewrite(address, PREFIX, 2, this->settings, this->size);
-    if (status != STORAGE_OK) {
+    if (status == STORAGE_OK) {
+    	status = storage->rewrite(address, PREFIX, 2, this->settings, this->size);
+    }
+	if (status != STORAGE_OK) {
 #if SETTINGS_BEDUG
         printTagLog(SettingsDB::TAG, "error save settings duplicate: storage save error=%02X address=%lu", status, address);
 #endif
         return SETTINGS_ERROR;
+    } else {
+#if SETTINGS_BEDUG
+        printTagLog(SettingsDB::TAG, "save duplicate settings: success address=%lu", address);
+#endif
     }
 
     if (this->load() == SETTINGS_OK) {
 #if SETTINGS_BEDUG
     	printTagLog(SettingsDB::TAG, "settings saved successfully (address=%lu)", address);
 #endif
-
     	return SETTINGS_OK;
     }