+++ Modules/at24cm01/at24cm01.c
@@ -12,8 +12,8 @@
 #include "utils.h"
 
 
-#define EEPROM_TIMER_DELAY_MS ((uint16_t)80)
-#define EEPROM_DELAY_MS       ((uint16_t)20)
+#define EEPROM_TIMER_DELAY_MS ((uint16_t)1000)
+#define EEPROM_DELAY_MS       ((uint16_t)100)
 #define EEPROM_MAX_ERRORS     (5)